Inform requesters that they likely forgot to add a path if they hit /
Webhooks from GitHub should be configured to hit
https://<your_url>/webhook by default. It can be easy to forget though,
and people may end up configuring https://<your_url>/ by mistake.

I found it easy to forget myself after not touching this repo for a
while. So not, instead of it saying "Hello, world!" on a GET, and a 405
Method Not Allowed on a POST, it will remind the user to check their
webhook path configuration :)